refactor: display name shared hook for both versions
