+++ src/ui/web/modules/settings/components/DisplayNameSection.tsx
@@ -1,13 +1,8 @@
-import { type ChangeEvent, useEffect, useMemo, useState } from "react";
+import type { ChangeEvent } from "react";
 
 import { CircleCheck, CircleX, Loader } from "lucide-react";
-import { useDebouncedCallback } from "use-debounce";
 
-import {
-  useCheckUsernameExists,
-  useCurrentUser,
-  useUpdateUsername,
-} from "@/ui/shared/user/hooks/useUser";
+import { useDisplayNameSection } from "@/ui/shared/settings/hooks/useDisplayNameSection";
 import { Small } from "@/ui/web/components/Typography";
 import { Button } from "@/ui/web/components/ui/button";
 import {
@@ -17,106 +12,26 @@ import {
 } from "@/ui/web/components/ui/input-group";
 
 export const DisplayNameSection = () => {
-  const currentUser = useCurrentUser();
-  const currentUsername = currentUser?.username?.trim() ?? "";
-  const [displayName, setDisplayName] = useState("");
-  const [hasEdited, setHasEdited] = useState(false);
-  const [doesNameExist, setDoesNameExist] = useState<boolean | null>(null);
-  const { checkUsernameExists, isChecking } = useCheckUsernameExists();
-  const updateUsername = useUpdateUsername();
-  const [isUpdating, setIsUpdating] = useState(false);
-
-  useEffect(() => {
-    if (hasEdited || displayName) {
-      return;
-    }
-
-    setDisplayName(currentUsername);
-    setDoesNameExist(currentUsername ? false : null);
-  }, [currentUsername, displayName, hasEdited]);
-
-  const trimmedDisplayName = useMemo(() => displayName.trim(), [displayName]);
-
-  const handleCheckUsername = async (value: string) => {
-    const trimmedValue = value.trim();
-
-    if (!trimmedValue) {
-      setDoesNameExist(null);
-      return;
-    }
-
-    if (trimmedValue === currentUsername) {
-      setDoesNameExist(false);
-      return;
-    }
-
-    try {
-      const exists = await checkUsernameExists(trimmedValue);
-      setDoesNameExist(exists);
-    } catch (error) {
-      console.error("Failed to verify username", error);
-      setDoesNameExist(null);
-    }
-  };
-
-  const debouncedCheckUsername = useDebouncedCallback((value: string) => {
-    void handleCheckUsername(value);
-  }, 600);
-
-  const handleChange = (event: ChangeEvent<HTMLInputElement>) => {
-    const nextValue = event.target.value;
-    setDisplayName(nextValue);
-    setHasEdited(true);
-
-    if (!nextValue.trim()) {
-      debouncedCheckUsername.cancel();
-      setDoesNameExist(null);
-      return;
-    }
-
-    if (nextValue.trim() === currentUsername) {
-      debouncedCheckUsername.cancel();
-      setDoesNameExist(false);
-      return;
-    }
-
-    debouncedCheckUsername(nextValue);
-  };
-
-  const handleAccept = async () => {
-    setIsUpdating(true);
-    try {
-      await updateUsername({ username: trimmedDisplayName });
-      setDisplayName(trimmedDisplayName);
-      setHasEdited(false);
-      setDoesNameExist(null);
-    } catch (error) {
-      console.error("Failed to update username", error);
-    } finally {
-      setIsUpdating(false);
-    }
-  };
-
-  const shouldShowAccept =
-    Boolean(trimmedDisplayName) &&
-    trimmedDisplayName !== currentUsername &&
-    doesNameExist === false &&
-    !isUpdating;
+  const {
+    displayName,
+    isBusy,
+    isUpdating,
+    onAcceptDisplayName,
+    onDisplayNameChanged,
+    shouldShowAccept,
+    status,
+  } = useDisplayNameSection();
 
   const renderStatusIcon = () => {
-    if (!trimmedDisplayName) {
-      return null;
-    }
-
-    if (isChecking || isUpdating) {
+    if (isBusy) {
       return <Loader className="size-6 animate-spin text-muted-foreground" />;
     }
 
-    if (doesNameExist === null) {
+    if (status === "empty" || status === "idle") {
       return null;
     }
 
-    return doesNameExist ? (
+    return status === "taken" ? (
       <CircleX className="size-6 text-destructive" />
     ) : (
       <CircleCheck className="size-6 text-primary" />
@@ -135,7 +50,7 @@ export const DisplayNameSection = () => {
             variant="outline"
             className={"text-primary tracking-widest"}
             size="sm"
-            onClick={() => void handleAccept()}
+            onClick={() => void onAcceptDisplayName()}
             disabled={isUpdating}
           >
             Accept
@@ -151,7 +66,9 @@ export const DisplayNameSection = () => {
             "text-base text-foreground placeholder:text-muted-foreground/70"
           }
           value={displayName}
-          onChange={handleChange}
+          onChange={(event: ChangeEvent<HTMLInputElement>) => {
+            onDisplayNameChanged(event.target.value);
+          }}
           disabled={isUpdating}
         />
         <InputGroupButton
